Implement deserialization for big arrays
The deserialization is partially unsafe since it writes into an
uninitialized array element by element.
The conversion from a map (in the serialized format) to an array of
tuples (in rust) is new and not available without const-generics
support.
